pyproject.toml
[project]
name = "tinyfacts-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jsonlines>=4.0.0",
    "lemminflect>=0.2.3",
    "nltk>=3.9.2",
    "openai>=1.99.6",
    "pydantic-ai==1.25.1",
    "python-dotenv>=1.1.1",
    "textual>=6.7.0",
    "typer>=0.20.0",
]
